fix(health-check): renamed files from appConfig Redux actions and reducer
