+++ public/redux/reducers/appConfigReducers.ts
