+++ public/redux/reducers/rootReducers.js
@@ -22,7 +22,7 @@ import visualizationsReducers from './visualizationsReducers';
 import globalBreadcrumbReducers from './globalBreadcrumbReducers';
 import securityReducers from './securityReducers';
 import toolsReducers from './toolsReducers';
-import appConfig from './app-config.reducer';
+import appConfig from './appConfigReducers';
 
 export default combineReducers({
   rulesetReducers,
